begin user command

+++ src/main/java/com/tsunyoku/osuplugin/models/BeatmapModel.java
@@ -6,15 +6,15 @@ public class BeatmapModel {
     public String DiffName;
     public String FileMD5;
     public int Mode;
-    public double BPM;
-    public double AR;
-    public double OD;
-    public double CS;
-    public double HP;
+    public float BPM;
+    public float AR;
+    public float OD;
+    public float CS;
+    public float HP;
     public int TotalLength;
     public int HitLength;
     public int Playcount;
     public int Passcount;
     public int MaxCombo;
-    public double DifficultyRating;
+    public float DifficultyRating;
 }
